+++ Source/Core/Image/MshColor.cpp
@@ -11,16 +11,13 @@
  *
  *  $Id$
  */
-/*----------------------------------------------------------------------------
- *  Reference:
- *  [1] Kenneth Moreland, "Diverging Color Maps for Scientific Visualization",
- *      In Proceedings of the 5th International Symposium on Visual Computing,
- *      December 2009. DOI 10.1007/978-3-642-10520-3_9.
- */
 /*****************************************************************************/
 #include "MshColor.h"
+#include "RGBColor.h"
 #include "LabColor.h"
+#include "XYZColor.h"
 #include <cmath>
+#include <kvs/Vector2>
 
 
 namespace
@@ -28,13 +25,28 @@ namespace
 
 /*===========================================================================*/
 /**
- *  @brief  Returns an adjusted hue when interpolating to un unsaturated color in Msh color space.
+ *  @brief  Returns angle difference in radian.
+ *  @param  rad1 [in] angle 1 [rad]
+ *  @param  rad2 [in] angle 2 [rad]
+ *  @return angle difference
+ */
+/*===========================================================================*/
+inline kvs::Real32 RadDiff( const kvs::Real32 rad1, const kvs::Real32 rad2 )
+{
+    const kvs::Vec2 v1( std::cos( rad1 ), std::sin( rad1 ) );
+    const kvs::Vec2 v2( std::cos( rad2 ), std::sin( rad2 ) );
+    return std::acos( v1.dot( v2 ) );
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Returns an adjusted hue when interpolating to an unsaturated color in Msh color space.
  *  @param  msh_st [in] saturated Msh color
  *  @param  M_ust [in] unsaturated M value of Msh color
  *  @return adjusted hue value
  */
 /*===========================================================================*/
-kvs::Real32 AdjustHue(
+inline kvs::Real32 AdjustHue(
     const kvs::MshColor& msh_st,
     const kvs::Real32 M_ust )
 {
@@ -42,7 +54,7 @@ kvs::Real32 AdjustHue(
     const kvs::Real32 s_st = msh_st.s();
     const kvs::Real32 h_st = msh_st.h();
 
-    if ( M_st >= M_ust )
+    if ( M_st >= M_ust - 0.1f )
     {
         return h_st;
     }
@@ -62,19 +74,19 @@ kvs::Real32 AdjustHue(
     }
 }
 
-kvs::MshColor Lab2Msh( const kvs::LabColor& lab )
+inline kvs::MshColor Lab2Msh( const kvs::LabColor& lab )
 {
     const kvs::Real32 l = lab.l();
     const kvs::Real32 a = lab.a();
     const kvs::Real32 b = lab.b();
 
     const kvs::Real32 m = std::sqrt( l * l + a * a + b * b );
-    const kvs::Real32 s = ( m > 0.00001 ) * std::acos( l / m );
-    const kvs::Real32 h = ( s > 0.00001 ) * std::atan2( b, a );
+    const kvs::Real32 s = ( m > 0.001f ) * std::acos( l / m );
+    const kvs::Real32 h = ( s > 0.001f ) * std::atan2( b, a );
     return kvs::MshColor( m, s, h );
 }
 
-kvs::LabColor Msh2Lab( const kvs::MshColor& msh )
+inline kvs::LabColor Msh2Lab( const kvs::MshColor& msh )
 {
     const kvs::Real32 m = msh.m();
     const kvs::Real32 s = msh.s();
@@ -84,6 +96,30 @@ kvs::LabColor Msh2Lab( const kvs::MshColor& msh )
     const kvs::Real32 a = m * std::sin( s ) * std::cos( h );
     const kvs::Real32 b = m * std::sin( s ) * std::sin( h );
     return kvs::LabColor( l, a, b );
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Converts color space from RGB to Msh.
+ *  @param  rgb [in] RGB color
+ *  @return Msh color
+ */
+/*===========================================================================*/
+inline kvs::MshColor RGB2Msh( const kvs::RGBColor& rgb )
+{
+    return rgb.toXYZColor().toLabColor().toMshColor();
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Converts color space from Msh to RGB.
+ *  @param  msh [in] Msh color
+ *  @return RGB color
+ */
+/*===========================================================================*/
+inline kvs::RGBColor Msh2RGB( const kvs::MshColor& msh )
+{
+    return msh.toLabColor().toXYZColor().toRGBColor();
 }
 
 }
@@ -105,7 +141,6 @@ kvs::MshColor MshColor::Mix(
     const kvs::MshColor& msh2,
     const kvs::Real32 t )
 {
-
     kvs::Real32 M1 = msh1.m();
     kvs::Real32 s1 = msh1.s();
     kvs::Real32 h1 = msh1.h();
@@ -115,10 +150,10 @@ kvs::MshColor MshColor::Mix(
     kvs::Real32 h2 = msh2.h();
 
     kvs::Real32 ratio = t;
-    if ( ( s1 > 0.05 ) && ( s2 > 0.05 ) && ( kvs::Math::Abs( h1 - h2 ) > kvs::Math::PI() / 3.0 ) )
+    if ( ( s1 > 0.05f ) && ( s2 > 0.05f ) && ( ::RadDiff( h1, h2 ) > kvs::Math::PI() / 3.0f ) )
     {
         const kvs::Real32 Mmid = kvs::Math::Max( M1, M2, 88.0f );
-        if ( t < 0.5 )
+        if ( t < 0.5f )
         {
             M2 = Mmid;
             s2 = 0.0;
@@ -132,14 +167,13 @@ kvs::MshColor MshColor::Mix(
             h1 = 0.0;
             ratio = 2.0f * ratio - 1.0f;
         }
-
     }
 
-    if ( ( s1 < 0.05 ) && ( s2 > 0.05 ) )
+    if ( ( s1 < 0.05f ) && ( s2 > 0.05f ) )
     {
         h1 = ::AdjustHue( kvs::MshColor( M2, s2, h2 ), M1 );
     }
-    else if ( ( s2 < 0.05 ) && ( s1 > 0.05 ) )
+    else if ( ( s2 < 0.05f ) && ( s1 > 0.05f ) )
     {
         h2 = ::AdjustHue( kvs::MshColor( M1, s1, h1 ), M2 );
     }
@@ -176,6 +210,16 @@ MshColor::MshColor( const kvs::LabColor& lab )
     *this = ::Lab2Msh( lab );
 }
 
+MshColor::MshColor( const kvs::RGBColor& rgb )
+{
+    *this = ::RGB2Msh( rgb );
+}
+
+kvs::RGBColor MshColor::toRGBColor() const
+{
+    return ::Msh2RGB( *this );
+}
+
 kvs::LabColor MshColor::toLabColor() const
 {
     return ::Msh2Lab( *this );
@@ -202,6 +246,12 @@ kvs::MshColor& MshColor::operator = ( const kvs::MshColor& msh )
     m_m = msh.m();
     m_s = msh.s();
     m_h = msh.h();
+    return *this;
+}
+
+kvs::MshColor& MshColor::operator = ( const kvs::RGBColor& rgb )
+{
+    *this = ::RGB2Msh( rgb );
     return *this;
 }